backend/pyproject.toml
[project]
name = "task-manager"
version = "0.1.0"
description = "Backend Manager App"
readme = "README.md"
authors = [{ name = "Tati Trif", email = "yatatitrif@yandex.com" }]
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.5.3",
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.38.0",
]
